Add files via upload
Hi I have added the 2nd nlp algorithm to extract the information and chat with the client.  Somehow I am not able to use the regex wisely. please try to improvise on that. So we get the best possible input. I am providing the training dataset text file as well.